requirements-bitri.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
docopt==0.6.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.9.0
indicnlp==0.0.1
joblib==1.1.0
Markdown==3.3.6
mosestokenizer==1.2.1
-e git+https://github.com/pegasus-lynx/nlcodec.git@ext-ngram#egg=nlcodec
numpy==1.21.4
oauthlib==3.1.1
openfile==0.0.7
portalocker==2.3.2
protobuf==3.19.1
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyspark==3.2.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
rtg==0.4.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sacrebleu-macrof==2.0.1
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
toolwrapper==2.1.0
torch==1.10.1
torchtext==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
uctools==1.3.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
